Add table of contents, docs on command line args and building for windows
The current readme lacks a table of contents, which makes finding what
you're looking for difficult because of the large amount of images and
text.

The current documentation on building for windows doesn't mention the
fact that package.cmd also has to be run to create a full build. It is
also lacking in other areas, such as the packaging programs required.

The current documentation doesn't reflect the fact that you can pass
command line arguments to the program.