Add compiler switches for warnings as errors
Currently with /W3 on MSVC there are no warnings. Clang and g++ are not
tested.

The "QMAKE_CFLAGS_WARN_ON -= -W3" and "QMAKE_CXXFLAGS_WARN_ON -= -W3"
are necessary because if they are not removed in the config, the
compiler will give D9025: overriding '/W4' with '/W3' error warnings and
not actually use /W4.