Fix shadowed variable in easylogging++.cc
The _var was chosen because it conflicts with the naming convention in
the rest of the file.

Fixing the error was easier than getting GCC to ignore errors in one
file, or using a pragma to mute the warning.
Introducing a GCC pragma would require disabling C4068 on MSVC (unknown
pragma).