Revert "Reduce update rate of chaperone velocity modifiers"
This reverts commit e712f4aa1f514d5875cb322a58f564362682612e.

The issue turned out to be isolated to Linux, since it wasn't possible
to reproduce the issue completely on Windows.

An MVP for the bug can be found here
https://github.com/username223/SteamVR-Manjaro-Sync-Crash

The issue is very likely to be on SteamVR's end, since the issue doesn't
appear on the same CPU running Windows.

The fix for the issue in
https://github.com/OpenVR-Advanced-Settings/OpenVR-AdvancedSettings/issues/252
is to by default turn off the offending feature. This doesn't fix
anything, but it will prevent people from having OVRAS kill their
performance just by running the application.