Added missing/changed options in (Application) settings page documentation
This change includes a fix for changes in naming for an option called "Disable Automatic Crash Recovery of Chaperone Config" that was changed in the application to "Enable..." likely due to OVRAS updates.

Missing options were added with a TBA placeholder for the details. I do not know the technical details of these options, and therefore advise to replace the TBA with something before merging to main.